third_party/blink/web_tests/external/wpt/editing/run/multitest_5001-6000-expected.txt

This is a testharness.js-based test.
Found 101 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["createlink","http://www.google.com/"],["justifyleft",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar" but got "foo<a href=\\"http://www.google.com/\\">http://www.google.com/</a>bar"
[FAIL] [["createlink","http://www.google.com/"],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["createlink","http://www.google.com/"],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["createlink","http://www.google.com/"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["createlink","http://www.google.com/"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["createlink","http://www.google.com/"],["justifyright",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foobar</div>" but got "<div style=\\"text-align:right\\">foo<a href=\\"http://www.google.com/\\">http://www.google.com/</a>bar</div>"
[FAIL] [["createlink","http://www.google.com/"],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["createlink","http://www.google.com/"],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["createlink","http://www.google.com/"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["createlink","http://www.google.com/"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["createlink","http://www.google.com/"],["outdent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar" but got "foo<a href=\\"http://www.google.com/\\">http://www.google.com/</a>bar"
[FAIL] [["fontname","sans-serif"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font face=\\"sans-serif\\">a</font>bar" but got "foo<span style=\\"font-family:sans-serif\\">a</span>bar"
[FAIL] [["fontname","sans-serif"],["delete",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<font face=\\"sans-serif\\">a</font>bar" but got "foabar"
[FAIL] [["fontname","sans-serif"],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["formatblock","<div>"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<font face=\\"sans-serif\\">a</font>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["fontname","sans-serif"],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["forwarddelete",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font face=\\"sans-serif\\">a</font>ar" but got "fooaar"
[FAIL] [["fontname","sans-serif"],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["fontname","sans-serif"],["indent",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<font face=\\"sans-serif\\">a</font>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["fontname","sans-serif"],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<hr><font face=\\"sans-serif\\">a</font>bar" but got "foo<hr>abar"
[FAIL] [["fontname","sans-serif"],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fooab<b>c</b>d<font face=\\"sans-serif\\">a</font>bar" but got "fooab<b>c</b>dabar"
[FAIL] [["fontname","sans-serif"],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<img src=\\"/img/lion.svg\\"><font face=\\"sans-serif\\">a</font>bar" but got "foo<img src=\\"/img/lion.svg\\">abar"
[FAIL] [["fontname","sans-serif"],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["insertlinebreak",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foo<span style=\\"font-family:sans-serif\\"><br></span>bar"
[FAIL] [["fontname","sans-serif"],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><font face=\\"sans-serif\\">a</font>bar" but got "foo<span style=\\"font-family:sans-serif\\"><br>a</span>bar"
[FAIL] [["fontname","sans-serif"],["insertorderedlist",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontname","sans-serif"],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontname","sans-serif"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<font face=\\"sans-serif\\">a</font>bar</li></ol>" but got "<ol><li>afoobar</li></ol>"
[FAIL] [["fontname","sans-serif"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontname","sans-serif"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontname","sans-serif"],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<div>bar</div>"
[FAIL] [["fontname","sans-serif"],["insertparagraph",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><font face=\\"sans-serif\\">a</font>bar</p>" but got "foo<div><span style=\\"font-family:sans-serif\\">a</span>bar</div>"
[FAIL] [["fontname","sans-serif"],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontname","sans-serif"],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontname","sans-serif"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<font face=\\"sans-serif\\">a</font>bar</li></ul>" but got "<ul><li>afoobar</li></ul>"
[FAIL] [["fontname","sans-serif"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontname","sans-serif"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontname","sans-serif"],["justifycenter",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontname","sans-serif"],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["fontname","sans-serif"],["justifycenter",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foo<font face=\\"sans-serif\\">a</font>bar</div>" but got "<div style=\\"text-align:center\\">fooabar</div>"
[FAIL] [["fontname","sans-serif"],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontname","sans-serif"],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["fontname","sans-serif"],["justifyfull",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontname","sans-serif"],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["fontname","sans-serif"],["justifyfull",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<font face=\\"sans-serif\\">a</font>bar</div>" but got "<div style=\\"text-align:justify\\">fooabar</div>"
[FAIL] [["fontname","sans-serif"],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontname","sans-serif"],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["fontname","sans-serif"],["justifyleft",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["fontname","sans-serif"],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["fontname","sans-serif"],["justifyleft",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font face=\\"sans-serif\\">a</font>bar" but got "fooabar"
[FAIL] [["fontname","sans-serif"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["fontname","sans-serif"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["fontname","sans-serif"],["justifyright",""]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontname","sans-serif"],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["fontname","sans-serif"],["justifyright",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foo<font face=\\"sans-serif\\">a</font>bar</div>" but got "<div style=\\"text-align:right\\">fooabar</div>"
[FAIL] [["fontname","sans-serif"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["fontname","sans-serif"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontname","sans-serif"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["fontname","sans-serif"],["outdent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font face=\\"sans-serif\\">a</font>bar" but got "foo<span style=\\"font-family:sans-serif\\">a</span>bar"
[FAIL] [["fontsize","4"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"4\\">a</font>bar" but got "foo<span style=\\"font-size:large\\">a</span>bar"
[FAIL] [["fontsize","4"],["delete",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<font size=\\"4\\">a</font>bar" but got "foabar"
[FAIL] [["fontsize","4"],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["formatblock","<div>"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<font size=\\"4\\">a</font>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["fontsize","4"],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["forwarddelete",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"4\\">a</font>ar" but got "fooaar"
[FAIL] [["fontsize","4"],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["fontsize","4"],["indent",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<font size=\\"4\\">a</font>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["fontsize","4"],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<hr><font size=\\"4\\">a</font>bar" but got "foo<hr>abar"
[FAIL] [["fontsize","4"],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fooab<b>c</b>d<font size=\\"4\\">a</font>bar" but got "fooab<b>c</b>dabar"
[FAIL] [["fontsize","4"],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
Harness: the test ran to completion.
